+++ src/main/java/ca/teamdman/sfml/ast/NumberRange.java
@@ -1,8 +1,8 @@
 package ca.teamdman.sfml.ast;
 
 public record NumberRange(
-        int start,
-        int end
+        long start,
+        long end
 ) implements ASTNode {
     public boolean contains(int value) {
         return value >= start && value <= end;
